feat(inventory): update Brand and Location exception and service implementations

+++ services/msvc-inventory/src/main/java/com/sssi/msvcinventory/service/impl/BrandServiceImpl.java
@@ -5,6 +5,7 @@ import com.sssi.msvcinventory.dto.response.BrandResponseDto;
 import com.sssi.msvcinventory.entity.Brand;
 import com.sssi.msvcinventory.exception.BrandException;
 import com.sssi.msvcinventory.mapper.BrandMapper;
+import com.sssi.msvcinventory.repository.AssetModelRepository;
 import com.sssi.msvcinventory.repository.BrandRepository;
 import com.sssi.msvcinventory.service.BrandService;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +21,7 @@ import java.util.UUID;
 public class BrandServiceImpl implements BrandService {
 
     private final BrandRepository brandRepository;
+    private final AssetModelRepository assetModelRepository;
     private final BrandMapper brandMapper;
 
     @Override
@@ -70,6 +72,10 @@ public class BrandServiceImpl implements BrandService {
 
         Brand brand = brandRepository.findById(id)
                 .orElseThrow(() -> BrandException.notFound(id.toString()));
+
+        if (assetModelRepository.existsByBrandId(id)) {
+            throw BrandException.inUse(id.toString());
+        }
 
         brandRepository.delete(brand);
     }
